pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "harvestkit"
version = "0.1.0"
description = "Universal job + business scraper. Python CLI engine + Chrome MV3 extension share a 60-field schema."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [{ name = "BluOryn" }]
keywords = ["scraper", "job-scraper", "ats", "schema.org", "playwright", "extension"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP :: Indexing/Search",
    "Topic :: Text Processing :: Markup :: HTML",
]
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.2.0",
    "PyYAML>=6.0.1",
    "tldextract>=5.1.2",
    "playwright>=1.45.0",
]

[project.optional-dependencies]
exports = ["gspread>=6.1.2", "google-auth>=2.33.0", "notion-client>=2.2.1"]
llm = ["anthropic>=0.42.0"]
dev = [
    "ruff>=0.4.0",
    "black>=24.4.0",
    "mypy>=1.10.0",
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "pre-commit>=3.7.0",
]

[project.urls]
Homepage = "https://github.com/BluOryn/HarvestKit"
Repository = "https://github.com/BluOryn/HarvestKit"
Issues = "https://github.com/BluOryn/HarvestKit/issues"

[project.scripts]
harvestkit = "job_scraper.main:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 110
target-version = "py310"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "N", "SIM", "C4"]
ignore = ["E501", "B008", "N803", "N806"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["B011"]
"src/job_scraper/extract.py" = ["E741"]  # ambiguous variable names allowed in regex maps

[tool.black]
line-length = 110
target-version = ["py310", "py311", "py312"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
warn_unused_ignores = true
strict_optional = false
files = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers"
markers = [
    "integration: marks tests that hit live external services",
    "playwright: marks tests that need a running browser",
]
